Cleanup: remove dead code, dedupe configs, inline single-use helpers
- Delete one-shot backfill_transcripts_to_events migration script
- Extract shared per-agent plugin config into stashai.plugin.agent_config;
  cursor/codex/gemini/openclaw/opencode configs now thin wrappers
  (claude-plugin untouched: it has a different env-var fast path)
- Inline _retry.with_retry into embeddings/__init__.py and move
  TransientEmbeddingError to embeddings/base.py; delete _retry.py
- Inline _lock_helpers.advisory_lock into session_summarizer worker;
  delete _lock_helpers.py
- Strip section-marker comments (// --- Constants ---, etc.) from
  frontend tables page
- Drop "Best-effort S3 cleanup" comment that restates `except: pass`
  in backend/routers/files.py

CustomSelect dedupe (between frontend/ and www/) and the users-router
device_name try/except deferred — see PR description.
