+++ backend/routers/files.py
@@ -319,7 +319,7 @@ async def delete_ws_file(
     try:
         await storage_service.delete_file(row["storage_key"])
     except Exception:
-        pass  # Best-effort S3 cleanup
+        pass
     await pool.execute(
         "DELETE FROM files WHERE id = $1 AND workspace_id = $2", file_id, workspace_id
     )
