+++ backend/services/embeddings/base.py
@@ -5,6 +5,14 @@ from abc import ABC, abstractmethod
 import numpy as np
 
 
+class TransientEmbeddingError(Exception):
+    """Provider call failed in a way that's worth retrying (429 / 5xx / network)."""
+
+    def __init__(self, message: str, retry_after: float | None = None):
+        super().__init__(message)
+        self.retry_after = retry_after
+
+
 class BaseEmbedder(ABC):
     """Override embed_batch() to bring your own embedding provider.
